Workflow build & move examples
