+++ Cargo.toml
@@ -1,11 +1,11 @@
 [workspace]
-members = ["hitori", "hitori-macros"]
+members = ["hitori", "hitori-examples", "hitori-macros"]
 
 [workspace.package]
-version = "0.2.2"
+version = "0.2.3"
 authors = ["Artur Helmanau <m30bit@gmail.com>"]
 license = "MIT OR Apache-2.0"
 edition = "2021"
 repository = "https://github.com/30bit/hitori"
-documentation = "https://docs.rs/hitori"
 rust-version = "1.56"
+readme = "../README.md"
