+++ README.md
@@ -4,7 +4,7 @@
 Hitori is generic compile-time regular expressions library. 
 It works by creating series of if-statements and for-loops for each expression. 
  
-*See code samples along with the traits, impls and structs they expand to in [`examples`].*
+*See code samples along with the traits, impls and structs they expand to in [examples].*
 
 # Limitations
 
@@ -23,7 +23,7 @@ large texts and an expression contains repetitions of frequent characters.
 - **`macros`** *(enabled by default)* – [`impl_expr_mut`] and [`impl_expr`] macros.
 - **`find-hitori`** – finds hitori package to be used in macros 
   even if it has been renamed in Cargo.toml. **`macros`** feature is required.
-- **`examples`** – includes [`examples`] module into the build.
+- **`examples`** – includes [examples] module into the build.
 
 # License
 
@@ -33,7 +33,7 @@ Unless you explicitly state otherwise, any contribution intentionally submitted
 for inclusion in this crate by you, as defined in the Apache-2.0 license, shall
 be dual licensed as above, without any additional terms or conditions.
 
-[`examples`]: https://docs.rs/hitori/latest/hitori/examples/index.html
+[examples]: https://docs.rs/hitori-examples
 [regex]: https://docs.rs/regex
 [hitori]: https://docs.rs/hitori
 [`impl_expr_mut`]: https://docs.rs/hitori/latest/hitori/attr.impl_expr.html
