+++ hitori-examples/src/all_patterns/hello.rs
@@ -1,0 +1,13 @@
+/// "hello"
+pub struct Hello;
+
+#[hitori::impl_expr]
+impl Expr<usize, char> for Hello {
+    const PATTERN: _ = (
+        |ch| ch == 'h',
+        |ch| ch == 'e',
+        |ch| ch == 'l',
+        |ch| ch == 'l',
+        |ch| ch == 'o',
+    );
+}
