+++ hitori-examples/src/all_patterns/true_.rs
@@ -1,0 +1,7 @@
+/// An empty all-pattern
+pub struct True;
+
+#[hitori::impl_expr]
+impl Expr<usize, char> for True {
+    const PATTERN: _ = ();
+}
