+++ hitori-examples/src/any_patterns/false_.rs
@@ -1,0 +1,7 @@
+/// An empty any-pattern
+pub struct False;
+
+#[hitori::impl_expr]
+impl Expr<usize, char> for False {
+    const PATTERN: _ = [];
+}
