+++ hitori-examples/src/lib.rs
@@ -1,0 +1,13 @@
+//! Structs and trait implementations generated by
+//! [`impl_expr`] and [`impl_expr_mut`] along with the source code.
+//!
+//! [`impl_expr`]: crate::impl_expr
+//! [`impl_expr_mut`]: crate::impl_expr
+
+pub mod all_patterns;
+pub mod any_patterns;
+pub mod capturing;
+pub mod generics;
+pub mod positions;
+pub mod putting_everything_together;
+pub mod repetitions;
