--- hitori/src/examples.rs
@@ -1,21 +1,0 @@
-//! Structs and trait implementations generated by
-//! [`impl_expr`] and [`impl_expr_mut`] along with the source code.
-//!
-//! [`impl_expr`]: crate::impl_expr
-//! [`impl_expr_mut`]: crate::impl_expr
-
-pub mod all_patterns;
-pub mod any_patterns;
-pub mod capturing;
-pub mod generics;
-pub mod positions;
-pub mod putting_everything_together;
-pub mod repetitions;
-
-macro_rules! include_example {
-    ($relative_path:literal) => {
-        concat!("# //", include_str!(concat!($relative_path, ".rs")))
-    };
-}
-
-pub(crate) use include_example;
