--- hitori/src/examples/all_patterns/hello.rs
@@ -1,14 +1,0 @@
-use crate as hitori;
-/// "hello"
-pub struct Hello;
-
-#[hitori::impl_expr]
-impl Expr<usize, char> for Hello {
-    const PATTERN: _ = (
-        |ch| ch == 'h',
-        |ch| ch == 'e',
-        |ch| ch == 'l',
-        |ch| ch == 'l',
-        |ch| ch == 'o',
-    );
-}
