--- hitori/src/examples/all_patterns/true_.rs
@@ -1,8 +1,0 @@
-use crate as hitori;
-/// An empty all-pattern
-pub struct True;
-
-#[hitori::impl_expr]
-impl Expr<usize, char> for True {
-    const PATTERN: _ = ();
-}
