--- hitori/src/examples/any_patterns/false_.rs
@@ -1,8 +1,0 @@
-use crate as hitori;
-/// An empty any-pattern
-pub struct False;
-
-#[hitori::impl_expr]
-impl Expr<usize, char> for False {
-    const PATTERN: _ = [];
-}
