+++ hitori/src/lib.rs
@@ -1,7 +1,7 @@
 //! Hitori is generic compile-time regular expressions library.
 //! It works by creating series of if-statements and for-loops for each expression.
 //!  
-//! *See code samples along with the traits, impls and structs they expand to in [`examples`].*
+//! *See code samples along with the traits, impls and structs they expand to in [examples].*
 //!
 //! # Limitations
 //!
@@ -20,8 +20,8 @@
 //! - **`macros`** *(enabled by default)* – [`impl_expr_mut`] and [`impl_expr`] macros.
 //! - **`find-hitori`** – finds hitori package to be used in macros
 //!   even if it has been renamed in Cargo.toml. **`macros`** feature is required.
-//! - **`examples`** – includes [`examples`] module into the build.
 //!
+//! [examples]: https://docs.rs/hitori-examples
 //! [hitori]: https://docs.rs/hitori
 //! [regex]: https://docs.rs/regex
 
@@ -32,7 +32,7 @@
     allow(mixed_script_confusables, confusable_idents)
 )]
 
-#[cfg(all(feature = "find-hitori", not(feature = "hitori-macros")))]
+#[cfg(all(feature = "find-hitori", not(feature = "macros")))]
 core::compile_error!(
     r#""find-hitori" feature doesn't do anything unless "macros" feature is enabled"#
 );
@@ -40,14 +40,6 @@ core::compile_error!(
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
-#[cfg(all(
-    any(doc, feature = "examples"),
-    feature = "alloc",
-    feature = "macros",
-    not(feature = "find_hitori")
-))]
-#[cfg_attr(doc, doc(cfg(feature = "examples")))]
-pub mod examples;
 pub mod string;
 
 mod expr;
